make DI helper functions discoverable by autoloader
